requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.11.7
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
dill==0.3.5.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
environs==9.5.0
frozenlist==1.3.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.4.1
marshmallow==3.17.0
mccabe==0.7.0
multidict==6.0.2
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pylint==2.14.5
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
wrapt==1.14.1
yarl==1.7.2
